requirements.txt
langchain
langchain-community
langchain-huggingface
langchain-google-genai
faiss-cpu
sentence-transformers
google-generativeai
python-dotenv
streamlit
fastapi
uvicorn
